restructured source folder for easier maintenance
