+++ src/bot/commands/index.js
