+++ src/bot/config.js
