+++ src/bot/control/index.js
