+++ src/bot/guards.js
