+++ src/bot/index.js
