+++ src/bot/random.js
