+++ src/bot/score.js
