+++ src/bot/utils/arrays.js
