+++ src/bot/utils/objects.js
