+++ src/bot/utils/strings.js
