+++ src/server/static/favicon.ico
