+++ test/unit/commands.js
@@ -1,9 +1,9 @@
 import { expect } from "chai";
 import sinon from "sinon";
-import { isAliveCommand, resetElection, setAccessCommand, setThrottleCommand, timetravelCommand } from "../../src/commands/commands.js";
-import { AccessLevel, CommandManager } from "../../src/commands/index.js";
-import Election from "../../src/election.js";
-import { dateToUtcTimestamp } from "../../src/utils/dates.js";
+import { isAliveCommand, resetElection, setAccessCommand, setThrottleCommand, timetravelCommand } from "../../src/bot/commands/commands.js";
+import { AccessLevel, CommandManager } from "../../src/bot/commands/index.js";
+import Election from "../../src/bot/election.js";
+import { dateToUtcTimestamp } from "../../src/bot/utils/dates.js";
 import { getMockBotConfig } from "../mocks/bot.js";
 import { getMockNominee } from "../mocks/nominee.js";
 import { getMockUserProfile } from "../mocks/user.js";
